미완 2차 수정 아직 JWT관련해서 좀 불편한 구석이 있음

+++ src/main/java/com/sparta/schedular_jpa/entity/UserSchedule.java
@@ -17,7 +17,7 @@ public class UserSchedule {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "user_schedule_id")
-    private int user_schedule_id;
+    private Long user_schedule_id;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "user_id")
